requirements.txt
# Core dependencies
typer>=0.9.0
rich>=13.7.0
srsly>=2.4.8
pathlib>=1.0.1

# Image processing
Pillow>=10.2.0
opencv-python>=4.9.0
numpy>=1.26.0
pdf2image>=1.17.0
pytesseract>=0.3.10

# Machine learning and data processing
scikit-learn>=1.4.0
torch>=2.2.0
transformers>=4.37.0

# Document processing
python-docx>=1.1.0

# Utilities
tqdm>=4.66.0
python-magic>=0.4.27 